feat: async trait
